添加微体检功能支持，2.2.60.15；
1.添加微体检功能支持；
2.优化JH58 PPG解析异常问题；

+++ iOS_sdk_source/Demo/VeepooBleSDKDemo/VeepooBleSDKDemo/VeepooBleSDK.framework/Headers/VPAccelerationModel.h
@@ -12,11 +12,11 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface VPAccelerationModel : NSObject
 
-@property(nonatomic, assign) double x;
+@property(nonatomic, assign) int16_t x;
 
-@property(nonatomic, assign) double y;
+@property(nonatomic, assign) int16_t y;
 
-@property(nonatomic, assign) double z;
+@property(nonatomic, assign) int16_t z;
 
 @end
